Added support for Firefox
There was a problem with Mozilla Firefox, because the Date it created was with a different reference. By forcing the Zulu hour, all browsers return the exact same Date object.

I've also changed the display so when the target date has been surpassed the count will show 0 (zero) in all the elements.